chore(ci): remove build and test stages from GitLab CI
This update eliminates the build and test stages from the GitLab CI configuration file. The decision to remove these stages may be due to changes in the workflow or migration of these processes to a different system. Additionally, the stage for 'changelog' has been updated to 'update' to better reflect its purpose in the pipeline.